Add authorizations

+++ src/Controllers/Admin/PostController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controllers\Admin;
 
+use App\Controllers\Auth\Auth;
 use App\Helpers\Request;
 use App\Helpers\DB;
 use App\Helpers\Router;
@@ -11,8 +12,19 @@ class PostController
 {
     public static function index(): void
     {
+        if (!Auth::check()) {
+            Router::redirect('home');
+        }
+
+        $query = 'SELECT p.id, p.title, p.slug, p.body, p.created_at, p.user_id
+        FROM posts AS p 
+        INNER JOIN users AS u 
+        ON p.user_id = u.id
+        WHERE p.user_id = ?';
+        
         $db = DB::connect();
-        $statement = $db->query('SELECT * FROM posts');
+        $statement = $db->prepare($query);
+        $statement->execute([Auth::id()]);
         $posts = $statement->fetchAll();
 
         require dirname(__DIR__, 3) . '/views/admin/posts/index.php';
@@ -32,43 +44,56 @@ class PostController
 
     public static function create(): void
     {
+        if (!Auth::check()) {
+            Router::redirect('home');
+        }
+
         require dirname(__DIR__, 3) . '/views/admin/posts/create.php';
-    }
+    }   
 
     public static function store(): void
     {
+        if (!Auth::check()) {
+            Router::redirect('home');
+        }
+
         $title = Request::input('title');
         $slug = Request::input('slug');
         $body = Request::input('body');
 
         if (empty($title)) {
-            Router::redirect('/admin/posts/create');
+            Router::redirect('admin.posts.create');
         }
 
         if (empty($slug)) {
-            Router::redirect('/admin/posts/create');
+            Router::redirect('admin.posts.create');
         }
 
         if (empty($body)) {
-            Router::redirect('/admin/posts/create');
+            Router::redirect('admin.posts.create');
         }
 
         try {
             $db = DB::connect();
-            $statement = $db->prepare('INSERT INTO posts (title, slug, body) VALUES (?,?,?)');
-            $statement->execute([$title, $slug, $body]);
+            $statement = $db->prepare('INSERT INTO posts (title, slug, body, user_id) VALUES (?,?,?,?)');
+            $statement->execute([$title, $slug, $body, Auth::id()]);
         } catch (PDOException) {
-            Router::redirect('/admin/posts/create');
+            Router::redirect('admin.posts.create');
         }
 
-        Router::redirect('/admin');
+        Router::redirect('admin');
     }
 
     public static function edit(int $id): void
     {
+        if (!Auth::check()) {
+            Router::redirect('home');
+        }
+
         $db = DB::connect();
-        $statement = $db->prepare('SELECT * FROM posts WHERE id = ?');
-        $statement->execute([$id]);
+        $statement = $db->prepare('SELECT * FROM posts WHERE id = ? AND user_id = ?');
+        $statement->execute([$id, Auth::id()]);
+
         $post = $statement->fetch();
 
         require dirname(__DIR__, 3) . '/views/admin/posts/edit.php';
@@ -76,35 +101,43 @@ class PostController
 
     public static function update(int $id): void
     {
+        if (!Auth::check()) {
+            Router::redirect('home');
+        }
+
         $title = Request::input('title');
         $slug = Request::input('slug');
         $body = Request::input('body');
 
         if (empty($title)) {
-            Router::redirect('/admin');
+            Router::redirect('admin');
         }
 
         if (empty($slug)) {
-            Router::redirect('/admin');
+            Router::redirect('admin');
         }
 
         if (empty($body)) {
-            Router::redirect('/admin');
+            Router::redirect('admin');
         }
 
         $db = DB::connect();
-        $statement = $db->prepare('UPDATE posts SET title = ?, slug = ?, body = ? WHERE id = ?');
-        $statement->execute([$title, $slug, $body, $id]);
+        $statement = $db->prepare('UPDATE posts SET title = ?, slug = ?, body = ?, user_id = ? WHERE id = ? AND user_id = ?');
+        $statement->execute([$title, $slug, $body, Auth::id(), $id, Auth::id()]);
 
-        Router::redirect('/admin');
+        Router::redirect('admin');
     }
 
     public static function delete(int $id): void
     {
+        if (!Auth::check()) {
+            Router::redirect('home');
+        }
+
         $db = DB::connect();
-        $statement = $db->prepare('DELETE FROM posts WHERE id = ?');
-        $statement->execute([$id]);
-        
-        Router::redirect('/admin');
+        $statement = $db->prepare('DELETE FROM posts WHERE id = ? AND user_id = ?');
+        $statement->execute([$id, Auth::id()]);
+
+        Router::redirect('admin');
     }
 }
